move udf tests to a separate file
